Introduce a browser session ID to overcome multiple CSR per visit
