+++ app/services/api.js
@@ -15,7 +15,6 @@ import {
 export default class ApiService extends Service {
   @service auth;
   @service fastboot;
-  @service visit;
 
   // Where the API lives
   // In production: https://rails.api.interflux.com
